refactor(GAME-039): extract hex geometry utilities to hex-geometry.ts
