+++ game/web/BUILD.bazel
@@ -24,6 +24,7 @@ ts_project(
         "//web/src/model:loader_lib",
         "//web/src/model:model_lib",
         "//web/src/model:types_lib",
+        "//web/src/model:hex_geometry_lib",
         "//web/src/model:generator_lib",
         "//web/src/model:adapter_lib",
         "//web/src/model:progress_lib",
@@ -54,6 +55,7 @@ esbuild(
         "//web/src/model:loader_lib",
         "//web/src/model:model_lib",
         "//web/src/model:types_lib",
+        "//web/src/model:hex_geometry_lib",
         "//web/src/model:generator_lib",
         "//web/src/model:adapter_lib",
         "//web/src/model:progress_lib",
